made the original test.py into testfile.py to differentiate between our testcase file and the file used for testing
